requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk
python-dotenv
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
oauth2client